requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==3.0.1